# i16->f32 a bit faster and more transparent