--- a/src/lib/xdsp/conv_f32_i16_2.c
+++ b/src/lib/xdsp/conv_f32_i16_2.c
@@ -18,11 +18,11 @@ VWLT_ATTRIBUTE(optimize("-O3"), target("sse2"))
 DECLARE_TR_FUNC_1_1(conv_f32_i16_sse2)
 #endif
 
-#ifdef WVLT_AVX
-#define TEMPLATE_FUNC_NAME conv_f32_i16_avx
-VWLT_ATTRIBUTE(optimize("-O3"), target("avx"))
-#include "templates/conv_f32_i16_generic.t"
-DECLARE_TR_FUNC_1_1(conv_f32_i16_avx)
+#ifdef WVLT_AVX2
+#define TEMPLATE_FUNC_NAME conv_f32_i16_avx2
+VWLT_ATTRIBUTE(optimize("-O3"), target("avx2"))
+#include "templates/conv_f32_i16_avx2.t"
+DECLARE_TR_FUNC_1_1(conv_f32_i16_avx2)
 #endif
 
 
@@ -34,7 +34,7 @@ conv_function_t conv_get_f32_i16_c(generic_opts_t cpu_cap, const char** sfunc)
 
     SELECT_GENERIC_FN(fn, fname, tr_conv_f32_i16_generic, cpu_cap);
     SELECT_SSE2_FN(fn, fname, tr_conv_f32_i16_sse2, cpu_cap);
-    SELECT_AVX_FN(fn, fname, tr_conv_f32_i16_avx, cpu_cap);
+    SELECT_AVX2_FN(fn, fname, tr_conv_f32_i16_avx2, cpu_cap);
 
     if (sfunc) *sfunc = fname;
     return fn;