--- a/src/lib/xdsp/utests/CMakeLists.txt
+++ b/src/lib/xdsp/utests/CMakeLists.txt
@@ -4,6 +4,7 @@
 set(TEST_SUIT_SRCS
     xdsp_utest_suite.c
     conv_i16_f32_utest.c
+    conv_f32_i16_utest.c
     conv_i12_f32_utest.c
     conv_ci12_2cf32_utest.c
     conv_f32_i12_utest.c
@@ -16,6 +17,7 @@ set(TEST_SUIT_SRCS
     ../fftad_functions.c
     ../rtsa_functions.c
     ../conv_i16_f32_2.c
+    ../conv_f32_i16_2.c
     ../conv_i12_f32_2.c
     ../conv_ci12_2cf32_2.c
     ../conv_f32_i12_2.c
@@ -24,7 +26,7 @@ set(TEST_SUIT_SRCS
 )
 
 if(WVLT_ARCH_X86 OR WVLT_ARCH_X86_64)
-    add_definitions("-mavx2 -mfma")
+    add_definitions("-msse4.1")
 endif()
 
 add_definitions("-O3")