--- a/src/lib/xdsp/utests/xdsp_utest_suite.c
+++ b/src/lib/xdsp/utests/xdsp_utest_suite.c
@@ -23,7 +23,7 @@ int main(int argc, char** argv)
 
     int number_failed;
     SRunner *sr;
-#if 1
+#if 0
     sr = srunner_create(conv_i16_f32_suite());
     srunner_add_suite(sr, fftad_suite());
     srunner_add_suite(sr, rtsa_suite());
@@ -33,7 +33,7 @@ int main(int argc, char** argv)
     srunner_add_suite(sr, conv_f32_i12_suite());
     srunner_add_suite(sr, conv_2cf32_ci12_suite());
 #else
-    sr = srunner_create(conv_ci12_2cf32_suite());
+    sr = srunner_create(conv_i16_f32_suite());
 #endif
     srunner_set_fork_status (sr, CK_NOFORK);
     srunner_run_all(sr, (argc > 1) ? CK_VERBOSE : CK_NORMAL);