--- a/src/lib/xdsp/utests/xdsp_utest_suite.c
+++ b/src/lib/xdsp/utests/xdsp_utest_suite.c
@@ -7,6 +7,7 @@
 #include "../vbase.h"
 
 Suite * conv_i16_f32_suite(void);
+Suite * conv_f32_i16_suite(void);
 Suite * fftad_suite(void);
 Suite * rtsa_suite(void);
 Suite * fft_window_cf32_suite(void);
@@ -25,6 +26,7 @@ int main(int argc, char** argv)
     SRunner *sr;
 #if 0
     sr = srunner_create(conv_i16_f32_suite());
+    srunner_add_suite(sr, conv_f32_i16_suite());
     srunner_add_suite(sr, fftad_suite());
     srunner_add_suite(sr, rtsa_suite());
     srunner_add_suite(sr, fft_window_cf32_suite());
@@ -33,7 +35,7 @@ int main(int argc, char** argv)
     srunner_add_suite(sr, conv_f32_i12_suite());
     srunner_add_suite(sr, conv_2cf32_ci12_suite());
 #else
-    sr = srunner_create(conv_i16_f32_suite());
+    sr = srunner_create(conv_f32_i16_suite());
 #endif
     srunner_set_fork_status (sr, CK_NOFORK);
     srunner_run_all(sr, (argc > 1) ? CK_VERBOSE : CK_NORMAL);