WAS
**** packet: 16384 bytes, iters: 1000000 ***
tr_conv_i16_f32_avx2		462054 us elapsed, 462 ns per 1 call, ave speed = 2164249 calls/s 
tr_conv_i16_f32_avx 		589653 us elapsed, 589 ns per 1 call, ave speed = 1695912 calls/s 
tr_conv_i16_f32_sse2		607927 us elapsed, 607 ns per 1 call, ave speed = 1644934 calls/s 
tr_conv_i16_f32_generic		2164669 us elapsed, 2164 ns per 1 call, ave speed = 461964 calls/s 

NOW
**** packet: 16384 bytes, iters: 1000000 ***
tr_conv_i16_f32_avx2		424700 us elapsed, 424 ns per 1 call, ave speed = 2354603 calls/s 
tr_conv_i16_f32_avx 		611548 us elapsed, 611 ns per 1 call, ave speed = 1635194 calls/s 
tr_conv_i16_f32_sse2		613781 us elapsed, 613 ns per 1 call, ave speed = 1629245 calls/s 
tr_conv_i16_f32_generic		2169088 us elapsed, 2169 ns per 1 call, ave speed = 461023 calls/s